requirements.txt
groq>=0.9.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
numpy>=1.26.0
python-dotenv>=1.0.0
